src/requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
channels==2.2.0
channels-redis==2.4.0
Django==2.2
django-cors-headers==2.5.3
django-redis==4.10.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
kombu==4.2.1
Pillow==5.4.0
psycopg2==2.7.3
PyJWT==1.7.1
pytz==2018.9
redis==3.2.0
rest-condition==1.0.3
vine==1.1.4
websockets==7.0